fixed build warnings
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS@4251 212acab6-be3b-0410-9dea-997c60f758d6

+++ trunk/Contrib/System/Source/Buffers.c
@@ -46,7 +46,7 @@ PLUGINFUNCTIONSHORT(Copy)
     if ((int) dest == 0) 
     {
         dest = GlobalAlloc((GPTR), size);
-        pushint(dest);
+        pushint((int) dest);
     }
 
     // COPY!
